# New implementation for tree-sitter based local detection