--- a/docker-images/syntax-highlighter/crates/scip-syntax/queries/matlab/scip-locals.scm
+++ b/docker-images/syntax-highlighter/crates/scip-syntax/queries/matlab/scip-locals.scm
@@ -1,12 +1,18 @@
+(function_definition) @scope.function
+(lambda) @scope
+(class_definition
+ name: (identifier) @definition.type
+ (#set! "hoist" "global")) @scope
+
 (assignment left: (identifier) @definition.var (#set! "reassignment_behavior" "oldest_is_definition"))
 (global_operator (identifier) @definition.var (#set! "reassignment_behavior" "oldest_is_definition"))
 (persistent_operator (identifier) @definition.var (#set! "reassignment_behavior" "oldest_is_definition"))
 
-(function_definition) @scope
 (function_definition
-    name: (identifier) @definition.function
-)
-(function_arguments 
+ name: (identifier) @definition.function
+ (#set! "hoist" "function"))
+
+(function_arguments
     (identifier) @definition.term
 )
 (function_output
@@ -18,9 +24,6 @@
     ]
 )
 
-(class_definition name: (identifier) @definition.type) @scope
-
-(lambda) @scope
 (lambda (arguments (identifier) @definition.term))
 
 (identifier) @reference