--- a/docker-images/syntax-highlighter/crates/scip-syntax/queries/matlab/scip-locals.scm
+++ b/docker-images/syntax-highlighter/crates/scip-syntax/queries/matlab/scip-locals.scm
@@ -7,23 +7,23 @@
 (assignment
  left: (identifier) @definition.var
  (#set! "reassignment_behavior" "oldest_is_definition")
- (#set! "is_def_ref"))
+ (#set! "def_ref"))
 (global_operator
  (identifier) @definition.var
  (#set! "reassignment_behavior" "oldest_is_definition")
- (#set! "is_def_ref"))
+ (#set! "def_ref"))
 (persistent_operator
  (identifier) @definition.var
  (#set! "reassignment_behavior" "oldest_is_definition")
- (#set! "is_def_ref"))
+ (#set! "def_ref"))
 
 (function_definition
  name: (identifier) @definition.function
  (#set! "hoist" "function"))
 
 (function_arguments
-    (identifier) @definition.term
-)
+ (identifier) @definition.term)
+
 (function_output
     [
         (multioutput_variable