--- a/docker-images/syntax-highlighter/crates/scip-syntax/src/bin/scip-perf.rs
+++ b/docker-images/syntax-highlighter/crates/scip-syntax/src/bin/scip-perf.rs
@@ -96,7 +96,7 @@ fn measure_parsing() {
     println!("Slowest files:");
     for timing in timings.iter().rev().take(10) {
         println!(
-            "{} ({}kb): {:?} ",
+            "{} ({}kb): {:?}",
             timing.file_path,
             timing.file_size / 1000,
             timing.duration