--- a/docker-images/syntax-highlighter/crates/scip-syntax/src/bin/scip-perf.rs
+++ b/docker-images/syntax-highlighter/crates/scip-syntax/src/bin/scip-perf.rs
@@ -2,7 +2,6 @@ use std::{path::Path, time::Instant};
 
 use clap::Parser;
 use scip_syntax::locals;
-use scip_syntax::locals_prev;
 use scip_treesitter_languages::parsers::BundledParser;
 use walkdir::WalkDir;
 
@@ -53,40 +52,6 @@ fn parse_files(dir: &Path) -> Vec<ParseTiming> {
     timings
 }
 
-fn parse_files_prev(dir: &Path) -> Vec<ParseTiming> {
-    let config = scip_syntax::languages::get_local_configuration(BundledParser::Go).unwrap();
-    let extension = "go";
-
-    let mut timings = vec![];
-
-    for entry in WalkDir::new(dir) {
-        let entry = entry.unwrap();
-        let entry = entry.path();
-
-        match entry.extension() {
-            Some(ext) if extension == ext => {}
-            _ => continue,
-        }
-
-        let start = Instant::now();
-        let source = std::fs::read_to_string(entry).unwrap();
-        let source_bytes = source.as_bytes();
-        let mut parser = config.get_parser();
-        let tree = parser.parse(source_bytes, None).unwrap();
-
-        locals_prev::parse_tree(config, &tree, source_bytes).unwrap();
-        let finish = Instant::now();
-
-        timings.push(ParseTiming {
-            filepath: entry.file_stem().unwrap().to_string_lossy().to_string(),
-            filesize: source_bytes.len(),
-            duration: finish - start,
-        });
-    }
-
-    timings
-}
-
 fn measure_parsing() {
     let args = Arguments::parse();
     println!("Measuring parsing");
@@ -111,46 +76,12 @@ fn measure_parsing() {
     println!("Done {:?}", finish - start);
 }
 
-fn measure_parsing_prev() {
-    let args = Arguments::parse();
-    println!("Measuring parsing");
-    let start = Instant::now();
-
-    let root = Path::new(&args.root_dir);
-
-    let mut timings = parse_files_prev(root);
-    timings.sort_by(|a, b| a.duration.cmp(&b.duration));
-    println!("Slowest files:");
-    for timing in timings.iter().rev().take(10) {
-        println!(
-            "{} ({}kb): {:?} ",
-            timing.filepath,
-            timing.filesize / 1000,
-            timing.duration
-        );
-    }
-
-    let finish = Instant::now();
-
-    println!("Done {:?}", finish - start);
-}
-
 fn main() {
     // TODO: parameterize
     let measure = "parsing";
 
     match measure {
-        "parsing" => {
-            println!("======");
-            println!("Previous");
-            println!("======");
-            measure_parsing_prev();
-
-            println!("======");
-            println!("New");
-            println!("======");
-            measure_parsing()
-        }
+        "parsing" => measure_parsing(),
         _ => panic!("Unknown measure: {}", measure),
     }
 }