… docker-images/syntax-highlighter/Cargo.lock
… docker-images/syntax-highlighter/crates/scip-syntax/Cargo.toml
… docker-images/syntax-highlighter/crates/scip-syntax/README.md
… docker-images/syntax-highlighter/crates/scip-syntax/README.org
… docker-images/syntax-highlighter/crates/scip-syntax/design/scoping.md
… docker-images/syntax-highlighter/crates/scip-syntax/docs/locals-query-dsl.md
… docker-images/syntax-highlighter/crates/scip-syntax/docs/locals-scoping.md
… docker-images/syntax-highlighter/crates/scip-syntax/src/locals_prev.rs
… docker-images/syntax-highlighter/crates/scip-syntax/src/snapshots/scip_syntax__locals__test__can_do_functions-2.snap
… docker-images/syntax-highlighter/crates/scip-syntax/src/snapshots/scip_syntax__locals__test__can_do_go-2.snap
… docker-images/syntax-highlighter/crates/scip-syntax/src/snapshots/scip_syntax__locals__test__can_do_java-2.snap
… docker-images/syntax-highlighter/crates/scip-syntax/src/snapshots/scip_syntax__locals__test__can_do_matlab-2.snap
… docker-images/syntax-highlighter/crates/scip-syntax/src/snapshots/scip_syntax__locals__test__can_do_matlab.snap
… docker-images/syntax-highlighter/crates/scip-syntax/src/snapshots/scip_syntax__locals__test__can_do_nested_locals-2.snap
… docker-images/syntax-highlighter/crates/scip-syntax/src/snapshots/scip_syntax__locals__test__can_do_perl-2.snap
… docker-images/syntax-highlighter/crates/scip-syntax/src/snapshots/scip_syntax__locals_prev__test__can_do_functions.snap
… docker-images/syntax-highlighter/crates/scip-syntax/src/snapshots/scip_syntax__locals_prev__test__can_do_go.snap
… docker-images/syntax-highlighter/crates/scip-syntax/src/snapshots/scip_syntax__locals_prev__test__can_do_java.snap
… docker-images/syntax-highlighter/crates/scip-syntax/src/snapshots/scip_syntax__locals_prev__test__can_do_matlab.snap
… docker-images/syntax-highlighter/crates/scip-syntax/src/snapshots/scip_syntax__locals_prev__test__can_do_nested_locals.snap
… docker-images/syntax-highlighter/crates/scip-syntax/src/snapshots/scip_syntax__locals_prev__test__can_do_perl.snap
… docker-images/syntax-highlighter/crates/scip-treesitter-cli/src/index.rs
… docker-images/syntax-highlighter/crates/scip-treesitter/src/lib.rs
… docker-images/syntax-highlighter/crates/sg-syntax/src/sg_treesitter.rs