## Description

A re-worked implementation of locals detection via tree-sitter queries.
The main difference to the previous implementation is that we now support lexical scoping and "targeted" hoisting.

### Lexical scoping

Let's look at a Go example to show the improved behaviour for lexical scoping:

```go
package main

var my_var = 1
func main() {
    my_var = 3 + my_var

    var my_var = 4
    my_var = 3 + my_var
}
```

The new implementation produces this diff when compared to the previous implementation:

```diff
  package main

  var my_var = 1
//    ^^^^^^ definition local 1
  func main() {
//     ^^^^ definition local 2
-      my_var = 3 + my_var
-//    ^^^^^^ reference local 3
-//                 ^^^^^^ reference local 3
+      my_var = 3 + my_var
+//    ^^^^^^ reference local 1
+//                 ^^^^^^ reference local 1

      var my_var = 4
//        ^^^^^^ definition local 3
      my_var = 3 + my_var
//    ^^^^^^ reference local 3
//                 ^^^^^^ reference local 3
```

Previously we'd indiscriminately lift all definitions to the top of their enclosing scope.
But Go's scoping rules for variables are purely lexical, so a definition is only in scope _after_ it was declared.
This means the first two references to `my_var` in this example have to reference the global definition.

Other languages that'll benefit from this include Perl, Rust, and OCaml (lol)

### Targeted hoisting

As we don't have local queries for JavaScript and Python yet I'll have to explain what this means without a diff in output.

Let's take this JavaScript example:
```js
function my_fun() {
  let x = 10;
  { var y = x };
  console.log(`${x} and ${y}`);
};
```

`var` is famously hoisted to the function level in JavaScript, and so both of the references in `console.log` should be resolved in this example.
Previously we'd only lift definitions to the closest enclosing scope, but with the new implementation we can specify that `var` bindings should be lifted to the closest enclosing scope of kind `function`.

A similar situtation exists for Python, where all local variables only exist at the function scope, and as such will need to be lifted as well.

## Test plan

My current best means of evaluating the new implementation against the previous one is to compare the test snapshots.
While definitions are renumbered, all references should still point to the "same" renumbered definition.

The new implementation also commits snapshots of the scope tree it generated during its execution. This should help us spot regressions in the future.

## Quick and dirty benchmarking

After a bit of profiling and fixing one egregious performance problem we are now slightly faster than the previous implementation for both Go and Java. Peak memory usage on the sourcegraph/sourcegraph repo also goes down from 150mb to 100mb. One thing of note is that the new implementation is significantly (2x) faster than the previous one on the largest Go file (3MB) in the sourcegraph repo.


<details>
<summary>Details</summary>

https://github.com/sourcegraph/sourcegraph

```
======
Previous
======
Measuring parsing
Slowest files:
mocks_temp (3319kb): 628.590667ms
mocks_test (704kb): 65.774708ms
mocks_test (430kb): 40.798334ms
mocks_test (440kb): 40.719875ms
mocks_test (427kb): 39.004792ms
gitserver.pb (247kb): 30.492917ms
mocks_test (303kb): 26.187458ms
mocks_temp (274kb): 25.060458ms
mocks_test (229kb): 19.002125ms
perms_store_test (129kb): 15.6135ms
Done 4.889227667s
======
New
======
Measuring parsing
Slowest files:
mocks_temp (3319kb): 350.291541ms
mocks_test (704kb): 54.395292ms
mocks_test (440kb): 35.303709ms
mocks_test (427kb): 35.22075ms
mocks_test (430kb): 34.949958ms
gitserver.pb (247kb): 28.288625ms
mocks_test (303kb): 23.554041ms
mocks_temp (274kb): 21.877375ms
mocks_test (229kb): 17.744042ms
perms_store_test (129kb): 15.049542ms
Done 4.491829541s

```

https://github.com/spring-projects/spring-framework:

```
======
Previous
======
Measuring parsing
Slowest files:
SpelCompilationCoverageTests (233kb): 23.238959ms
ServletAnnotationControllerHandlerMethodTests (170kb): 15.837125ms
AutowiredAnnotationBeanPostProcessorTests (163kb): 15.81025ms
DefaultListableBeanFactoryTests (132kb): 13.253625ms
MergedAnnotationsTests (144kb): 11.031292ms
DataBinderTests (92kb): 9.773125ms
ClassReader (172kb): 9.5185ms
SpelReproTests (79kb): 7.625125ms
AbstractPropertyAccessorTests (71kb): 7.518709ms
ResolvableTypeTests (70kb): 6.893208ms
Done 3.4509975s
======
New
======
Measuring parsing
Slowest files:
SpelCompilationCoverageTests (233kb): 23.672458ms
AutowiredAnnotationBeanPostProcessorTests (163kb): 15.413667ms
ServletAnnotationControllerHandlerMethodTests (170kb): 14.913167ms
DefaultListableBeanFactoryTests (132kb): 12.959834ms
MergedAnnotationsTests (144kb): 10.826292ms
DataBinderTests (92kb): 9.529833ms
ClassReader (172kb): 9.451625ms
AbstractPropertyAccessorTests (71kb): 7.398167ms
SpelReproTests (79kb): 7.373459ms
ResolvableTypeTests (70kb): 6.700875ms
Done 3.368848875s
```

https://github.com/aludnam/MATLAB:

```
======
Previous
======
Measuring parsing
Slowest files:
mtimesx_test_ssspeed (420kb): 21.280625ms
mtimesx_test_sdspeed (393kb): 19.575ms
mtimesx_test_dsspeed (392kb): 19.428042ms
mtimesx_test_ddspeed (121kb): 17.436875ms
mtimesx_test_ssequal (359kb): 16.542208ms
mtimesx_test_sdequal (354kb): 16.2475ms
mtimesx_test_dsequal (354kb): 15.889459ms
mtimesx_test_ddequal (94kb): 12.933917ms
gui_cp (80kb): 9.220958ms
nmf_smart_alg (23kb): 5.752375ms
Done 934.253625ms
======
New
======
Measuring parsing
Slowest files:
mtimesx_test_ssspeed (420kb): 22.3835ms
mtimesx_test_sdspeed (393kb): 20.55ms
mtimesx_test_dsspeed (392kb): 20.250375ms
mtimesx_test_ddspeed (121kb): 18.290083ms
mtimesx_test_ssequal (359kb): 17.372833ms
mtimesx_test_sdequal (354kb): 16.939709ms
mtimesx_test_dsequal (354kb): 16.5ms
mtimesx_test_ddequal (94kb): 13.711ms
gui_cp (80kb): 8.552125ms
exportfig (33kb): 3.653959ms
Done 921.589833ms
```

</details>