@kritzcreek 2023-12-22
I've done some performance testing against a couple large-ish Matlab projects on GitHub and the new implementation is slightly faster for all of them. I added a run against the largest repo I could find to the initial description.